GH Issue #268 Version Check Too Strict
When checking the version of the from module in task.ps1, the version
check is very strict. A difference in the z release of the installed
module will cause the module to be rejected and the task to fail.

This change loosen's the criteria of the check so that the Major version
is ensured to be the same, and Minor version is ensured to be equal to
or greater. Patch release version is ignored.